Fixed Gui, packages and added some libraries
Radiobuttons are not longer used in the proyect, instead i use ComboBoxes in order to make an easier use of the app, also, it will avoid issues with exceptions.

The classes now are in their correct package.

Added Material library, swing material library, JCore library and a self made library for gui components( and the oceanic look and feel).